Deleted duplicated Resource Manager, deleted state in Resource manager
